refactor: changed facet and diamond addresses
