+++ simulations/vip-diamond-comptroller/scripts/diamond.ts
