+++ simulations/vip-diamond-comptroller/scripts/facet-cut-params-generator.ts
