+++ simulations/vip-diamond-comptroller/vip-diamond-comptroller-testnet/utils/cut-params-testnet.json
@@ -1,7 +1,7 @@
 {
   "cutParams": [
     [
-      "0xBa70F1c3Fe2922992843Aa429DdE9d9Bfa163360",
+      "0x1B9806d9d2925e8Cd318E268e562eeb7e02C6E00",
       0,
       [
         "0xa76b3fda",
@@ -17,7 +17,7 @@
       ]
     ],
     [
-      "0x12eB32007a640eB71B5Fe412ea774CfB186aB220",
+      "0xe13A002aff3c64A72F0b9eD72d7fd861C5E0F280",
       0,
       [
         "0xead1a8a0",
@@ -40,7 +40,7 @@
       ]
     ],
     [
-      "0x34F3e6A812DaaB15803Aa9C38A22699d1Dfdc88A",
+      "0x19fB43e4b46DCB803208253E42F5D1800D2ACEca",
       0,
       [
         "0xa7604b41",
@@ -55,7 +55,7 @@
       ]
     ],
     [
-      "0x144301235b35811C7eb7a838565F842766AE6B49",
+      "0x68f0C17bf638AB86024B923c643fD15d3A6FC28B",
       0,
       [
         "0xf519fc30",
