--- simulations/vip-diamond-comptroller/vip-diamond-comptroller/abi/priceOracleAbi.json
@@ -1,154 +1,0 @@
-[
-  {
-    "inputs": [{ "internalType": "uint256", "name": "maxStalePeriod_", "type": "uint256" }],
-    "payable": false,
-    "stateMutability": "nonpayable",
-    "type": "constructor"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      { "indexed": false, "internalType": "address", "name": "feed", "type": "address" },
-      { "indexed": false, "internalType": "string", "name": "symbol", "type": "string" }
-    ],
-    "name": "FeedSet",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      { "indexed": false, "internalType": "uint256", "name": "oldMaxStalePeriod", "type": "uint256" },
-      { "indexed": false, "internalType": "uint256", "name": "newMaxStalePeriod", "type": "uint256" }
-    ],
-    "name": "MaxStalePeriodUpdated",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      { "indexed": false, "internalType": "address", "name": "oldAdmin", "type": "address" },
-      { "indexed": false, "internalType": "address", "name": "newAdmin", "type": "address" }
-    ],
-    "name": "NewAdmin",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      { "indexed": false, "internalType": "address", "name": "asset", "type": "address" },
-      { "indexed": false, "internalType": "uint256", "name": "previousPriceMantissa", "type": "uint256" },
-      { "indexed": false, "internalType": "uint256", "name": "requestedPriceMantissa", "type": "uint256" },
-      { "indexed": false, "internalType": "uint256", "name": "newPriceMantissa", "type": "uint256" }
-    ],
-    "name": "PricePosted",
-    "type": "event"
-  },
-  {
-    "constant": true,
-    "inputs": [],
-    "name": "admin",
-    "outputs": [{ "internalType": "address", "name": "", "type": "address" }],
-    "payable": false,
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [{ "internalType": "address", "name": "asset", "type": "address" }],
-    "name": "assetPrices",
-    "outputs": [{ "internalType": "uint256", "name": "", "type": "uint256" }],
-    "payable": false,
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [{ "internalType": "string", "name": "symbol", "type": "string" }],
-    "name": "getFeed",
-    "outputs": [{ "internalType": "contract AggregatorV2V3Interface", "name": "", "type": "address" }],
-    "payable": false,
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [{ "internalType": "contract VToken", "name": "vToken", "type": "address" }],
-    "name": "getUnderlyingPrice",
-    "outputs": [{ "internalType": "uint256", "name": "", "type": "uint256" }],
-    "payable": false,
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [],
-    "name": "isPriceOracle",
-    "outputs": [{ "internalType": "bool", "name": "", "type": "bool" }],
-    "payable": false,
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [],
-    "name": "maxStalePeriod",
-    "outputs": [{ "internalType": "uint256", "name": "", "type": "uint256" }],
-    "payable": false,
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [{ "internalType": "address", "name": "newAdmin", "type": "address" }],
-    "name": "setAdmin",
-    "outputs": [],
-    "payable": false,
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      { "internalType": "address", "name": "asset", "type": "address" },
-      { "internalType": "uint256", "name": "price", "type": "uint256" }
-    ],
-    "name": "setDirectPrice",
-    "outputs": [],
-    "payable": false,
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      { "internalType": "string", "name": "symbol", "type": "string" },
-      { "internalType": "address", "name": "feed", "type": "address" }
-    ],
-    "name": "setFeed",
-    "outputs": [],
-    "payable": false,
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [{ "internalType": "uint256", "name": "newMaxStalePeriod", "type": "uint256" }],
-    "name": "setMaxStalePeriod",
-    "outputs": [],
-    "payable": false,
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      { "internalType": "contract VToken", "name": "vToken", "type": "address" },
-      { "internalType": "uint256", "name": "underlyingPriceMantissa", "type": "uint256" }
-    ],
-    "name": "setUnderlyingPrice",
-    "outputs": [],
-    "payable": false,
-    "stateMutability": "nonpayable",
-    "type": "function"
-  }
-]
